// types/filters.ts
// types/filters.ts

export type Filters = {
  q?: string;                     // metin arama
  category?: string[];            // Üst, Alt, Ayakkabı...
  color?: string[];               // Mavi, Siyah...
  brand?: string[];               // Markalar
  season?: string[];              // İlkbahar, Yaz...
  dress_code?: string[];          // Casual, Business Casual...
  tags?: string[];                // Örn: Vintage, Wedding...
  fit?: string[];                 // Dar / Normal / Bol
  sort?: "newest" | "oldest" | "az" | "za";
};